+++ src/finalizer/index.ts
@@ -4,12 +4,10 @@ import {
   startupLogLevel,
   processEndPollingLoop,
   Contract,
-  ethers,
   getNetworkName,
   etherscanLink,
   getBlockForTimestamp,
   getCurrentTime,
-  getRedis,
   disconnectRedisClient,
 } from "../utils";
 import { winston } from "../utils";
@@ -30,17 +28,14 @@ import {
   Clients,
   ProcessEnv,
   FINALIZER_TOKENBRIDGE_LOOKBACK,
+  multicall3Addresses,
+  Multicall2Call,
 } from "../common";
-import { Multicall2Ethers__factory } from "@uma/contracts-node";
+import { getAbi } from "@uma/contracts-node";
 import * as optimismSDK from "@eth-optimism/sdk";
 import * as bobaSDK from "@across-protocol/boba-sdk";
 config();
 let logger: winston.Logger;
-
-export interface Multicall2Call {
-  callData: ethers.utils.BytesLike;
-  target: string;
-}
 
 export interface Withdrawal {
   l2ChainId: number;
@@ -63,11 +58,7 @@ export async function finalize(
   const hubPoolClient = configStoreClient.hubPoolClient;
   // Note: Could move this into a client in the future to manage # of calls and chunk calls based on
   // input byte length.
-  const multicall2 = new Contract(
-    "0x5ba1e12693dc8f9c48aad8770482f4739beed696",
-    Multicall2Ethers__factory.abi,
-    hubPoolClient.hubPool.signer
-  );
+  const multicall2 = new Contract(multicall3Addresses[1], getAbi("Multicall2"), hubPoolClient.hubPool.signer);
   const finalizationsToBatch: { callData: Multicall2Call[]; withdrawals: Withdrawal[] } = {
     callData: [],
     withdrawals: [],
